Filter on committeess page
